+++ pykechain/__about__.py
@@ -2,7 +2,7 @@
 name = 'pykechain'
 description = 'KE-chain Python SDK'
 
-version = '2.5.3'
+version = '2.6.0'
 
 author = 'KE-works BV'
 email = 'support+pykechain@ke-works.com'